generic simulation from reactor
